\name{additive}
\alias{additve}
\title{Incomplete Distance Matrix Filling}
\usage{additive(X)}
\arguments{\item{X}{a distance matrix}}
\description{Fills missing entries from incomplete distance matrix using the additive procedure (see reference for details)}
\value{
a distance matrix
}
\references{A weighted least-squares approach for inferring phylogenies from incomplete distance matrices
Vladimir Makarenkov and François-Joseph Lapointe, Vol. 20 no. 13 2004, pages 2113–2121
BIOINFORMATICS doi:10.1093/bioinformatics/bth211}
\author{C code and interfacing with \R by Andrei Popescu \email{niteloserpopescu@gmail.com}}