% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{AP}
\alias{AP}
\title{Adaptive Projection Estimation}
\usage{
AP(X, y, S, alpha = 0.05)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response}

\item{S}{the prescreened set}

\item{alpha}{the significance level alpha, default is 0.05}
}
\value{
An APE object is returned, including confidence intervals for regression coefficients.
}
\description{
Adaptive Projection Estimation
}
\examples{
\dontrun{
dat <- simdata(n, p, rho, truebeta)
X <- dat$X
y <- dat$y
S <- SIS(X, y, family = 'gaussian', tune = 'bic', seed = 2)$sis.ix0
ape <- AP(X, y, S)
CI <- ape$CI; len_CI <- ape$len_CI
}
}
