\name{ape-internal}
\alias{perm.rowscols}
\alias{phylogram.plot}
\alias{cladogram.plot}
\alias{circular.plot}
\alias{unrooted.plot}
\alias{unrooted.xy}
\alias{BOTHlabels}
\alias{floating.pie.asp}
\alias{plotPhyloCoor}
\alias{postprocess.prop.part}
\alias{ONEwise}
\alias{reorderRcpp}
\alias{polar2rect}
\alias{rect2polar}
\alias{neworder_phylo}
\alias{neworder_pruningwise}
\alias{node_depth}
\alias{node_depth_edgelength}
\alias{node_height}
\alias{node_height_clado}
\alias{seq_root2tip}
\title{Internal Ape Functions}
\description{
  Internal \pkg{ape} functions which are undocumented but still exported
  because called by other packages. Use with care!
}
\keyword{internal}
