\name{arecompatible.quartet}
\alias{arecompatible.quartet}
\title{Method for checking if a set of quartets is compatible}
\usage{arecompatible.quartet(x)}
\arguments{\item{x}{an object of class quartet}}
\description{Checks whether a set of quartets is compatible (see reference for details)}
\value{
TRUE if the quartets are compatible, FALSE otherwise
}
\references{Encoding phylogenetic trees in terms of weighted quartets. S.Gruenewald, K.T. Huber V. Moulton, C. Semple. Journal of Mathematical Biology 56(4) (2008) 465-477
}
\author{Andrei Popescu \email{niteloserpopescu@gmail.com}}