\name{arecompatible}
\alias{arecompatible}
\title{Method for checking if two splits modelled as in the splits class are compatible}
\usage{arecompatible(x,y,n)}
\arguments{\item{x}{an array of raw}\item{y}{an array of raw}\item{n}{the number of taxa in the splits}}
\description{Checks whether two splits are compatible}
\value{
TRUE if the splits are compatible, FALSE otherwise
}
\author{Andrei Popescu \email{niteloserpopescu@gmail.com}}