\name{data.nex}
\docType{data}
\alias{data.nex}
\alias{cynipids}
\title{NEXUS Data Example}
\description{
    Example of Protein data in NEXUS format (Maddison et al., 1997).
    Data is written in interleaved format using a single DATA block.
    Original data from Rokas et al (2002).
}
\usage{data(cynipids)}
\format{ASCII text in NEXUS format}
\references{
  Maddison, D. R., Swofford, D. L. and Maddison, W. P. (1997) NEXUS: an
  extensible file format for systematic information. \emph{Systematic
    Biology}, \bold{46}, 590--621.

  Rokas, A., Nylander, J. A. A., Ronquist, F. and Stone, G. N. (2002) A
  maximum likelihood analysis of eight phylogenetic markers in Gallwasps
  (Hymenoptera: Cynipidae): implications for insect phylogenetic
  studies. \emph{Molecular Phylogenetics and Evolution}, \bold{22},
  206--219.
}
\keyword{datasets}
