\name{hasNames}
\alias{hasNames}
\title{Check for names in a vector, matrix or data.frame}
\usage{hasNames(x)}
\arguments{
  \item{x}{a vector, matrix or data.frame}
}
\description{
  \code{hasNames} is a utility function for checking if a vector, matrix or data.frame has names or row.names.
}
\value{
 TRUE if names are present, FALSE otherwise
}
\author{Steven Kembel \email{skembel@berkeley.edu}}

\keyword{manip}
