\name{iscompatible}
\alias{iscompatible}
\title{Method for checking a split system modelled by the splits class is compatible}
\usage{iscompatible(x)}
\arguments{\item{x}{a splits object}}
\description{Checks whether a set of splits is compatible using the arecompatible method}
\value{
TRUE if the splits are compatible, FALSE otherwise
}
\author{Andrei Popescu \email{niteloserpopescu@gmail.com}}