\name{mat3}
\alias{mat3}
\title{Three Matrices}
\description{
  Three matrices respectively representing Serological (asymmetric),
  DNA hybridization (asymmetric) and Anatomical (symmetric) distances
  among 9 families.
}
\usage{
data(mat3)
}
\format{
  A data frame with 27 observations and 9 variables.
}
\source{
  Lapointe, F.-J., J. A. W. Kirsch and J. M. Hutcheon. 1999. Total
  evidence, consensus, and bat phylogeny: a distance-based
  approach. Molecular Phylogenetics and Evolution 11: 55-66.
}
\seealso{
  \code{\link{mat5Mrand}}, \code{\link{mat5M3ID}}
}
\keyword{datasets}
