\name{mat5M3ID}
\alias{mat5M3ID}
\title{Five Trees}
\description{
  Three partly similar trees, two independent trees.
}
\usage{
data(mat5M3ID)
}
\format{
  A data frame with 250 observations and 50 variables.
}
\source{
  Data provided by V. Campbell.
}
\seealso{
  \code{\link{mat5Mrand}}, \code{\link{mat3}}
}
\keyword{datasets}
