\name{mat5Mrand}
\alias{mat5Mrand}
\title{Five Independent Trees}
\description{
  Five independent additive trees.
}
\usage{
data(mat5Mrand)
}
\format{
  A data frame with 250 observations and 50 variables.
}
\source{
  Data provided by V. Campbell.
}
\seealso{
  \code{\link{mat5M3ID}}, \code{\link{mat3}}
}
\keyword{datasets}
