\name{StatComp21003-package}
\alias{StatComp21003-package}
\alias{StatComp21003}
\docType{package}
\title{
  High Dimensional statistical Inference
}
\description{
  Algorithms for adaptive projection estimation and low dimensional projection estimation under high dimensional linear model. And all homeworks in stat comp course.
}
\author{
  Hangjun Guo
  
  Maintainer: Hangjun Guo <hangjunguo@mail.ustc.edu.cn>
}
\references{
  ZHANG C H, ZHANG S S. Confidence intervals for low dimensional parameters in high dimensional linear models[J]. Journal of the Royal Statistical Society: Series B (Statistical Methodology),2014:217-242.
  
  ZHENG Z, ZHOU J, LI Y. Large-scale inference via the adaptive projection estimator. Unpublished Manuscript.
}
\keyword{ package }
\examples{
  ## see examples in AP and LDP  
}
