\name{treePop}
\alias{treePop}
\title{Method for reconstructing phylogenetic trees from an object of class splits}
\usage{treePop(x)}
\arguments{\item{x}{a splits object}}
\description{Method for reconstructing phylogenetic trees from an object of class splits using tree popping}
\value{an object of class "phylo" which displays all the splits in the input splits object}
\author{Andrei Popescu \email{niteloserpopescu@gmail.com}}