\name{ape-internal}
\alias{tree.build}
\alias{f.cherry.yule}
\alias{f.cherry.uniform}
\alias{sortIndex}
\alias{nsca}
\alias{perm.rowscols}
\alias{mant.zstat}
\alias{lower.triang}
\alias{clado.build}
\alias{phylogram.plot}
\alias{cladogram.plot}
\alias{circular.plot}
\alias{unrooted.plot}
\alias{unrooted.xy}
\alias{birth.step}
\alias{death.step}
\alias{ht.move}
\alias{loglik.pop}
\alias{pos.move}
\alias{BOTHlabels}
\alias{chronopl.cv}
\alias{floating.pie.asp}
\alias{checkLabel}
\alias{plotCophylo2}
\alias{plotPhyloCoor}
\alias{integrateTrapeze}
\alias{CDF.birth.death}
\alias{postprocess.prop.part}
\alias{ONEwise}
\title{Internal Ape Functions}
\description{
  Internal \pkg{ape} functions.
}
\keyword{internal}
