\name{ape-internal}
\alias{tree.build}
\alias{f.cherry.yule}
\alias{f.cherry.uniform}
\alias{sort.index}
\alias{nsca}
\alias{get.dna}
\alias{perm.rowscols}
\alias{mant.zstat}
\alias{lower.triang}
\alias{clado.build}
\alias{getDurations}
\alias{getEdgeLengths}
\alias{getExternalParams}
\alias{getNEdges}
\alias{getNFreeParams}
\alias{getRates}
\alias{objFuncLogScale}
\alias{optimTree}
\alias{phylogram}
\alias{prepareTree}
\alias{setTree}
\alias{buildTreeFromHclust}
\alias{buildTreeFromPhylo}
\alias{destroyTree}
\alias{edgeLengthsForPhylo}
\alias{getError}
\alias{getHclust}
\alias{getPhylo}
\alias{leftNodesForHclust}
\alias{lowerNodesForPhylo}
\alias{nEdges}
\alias{nNodes}
\alias{nodeHeightsForHclust}
\alias{nTips}
\alias{rightNodesForHclust}
\alias{tipLabelsForPhylo}
\alias{upperNodesForPhylo}
\alias{getMisawaTajima}
\title{Internal Ape Functions}
\description{
  Internal ape functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
