\name{ape-internal}
\alias{tree.build}
\alias{f.cherry.yule}
\alias{f.cherry.uniform}
\alias{sortIndex}
\alias{nsca}
\alias{perm.rowscols}
\alias{mant.zstat}
\alias{lower.triang}
\alias{clado.build}
\alias{phylogram.plot}
\alias{cladogram.plot}
\alias{circular.plot}
\alias{unrooted.plot}
\alias{unrooted.xy}
\alias{birth.step}
\alias{death.step}
\alias{ht.move}
\alias{loglik.pop}
\alias{pos.move}
\alias{BOTHlabels}
\alias{chronopl.cv}
\alias{floating.pie.asp}
\alias{checkLabel}
\alias{plotCophylo2}
\alias{plotPhyloCoor}
\alias{integrateTrapeze}
\alias{CDF.birth.death}
\alias{postprocess.prop.part}
\alias{ONEwise}
\alias{BaseProportion}
\alias{C_additive}
\alias{C_bionj}
\alias{C_bionjs}
\alias{C_ewLasso}
\alias{C_mvr}
\alias{C_mvrs}
\alias{C_nj}
\alias{C_njs}
\alias{C_pic}
\alias{C_rTraitCont}
\alias{C_treePop}
\alias{C_triangMtd}
\alias{C_triangMtds}
\alias{C_ultrametric}
\alias{C_where}
\alias{GlobalDeletionDNA}
\alias{SegSites}
\alias{bipartition}
\alias{delta_plot}
\alias{dist_dna}
\alias{dist_nodes}
\alias{mat_expo}
\alias{me_b}
\alias{me_o}
\alias{neworder_phylo}
\alias{neworder_pruningwise}
\alias{node_depth}
\alias{node_depth_edgelength}
\alias{node_height}
\alias{node_height_clado}
\alias{prop_part}
\alias{rawStreamToDNAbin}
\alias{seq_root2tip}
\alias{treeBuildWithTokens}
\title{Internal Ape Functions}
\description{
  Internal \pkg{ape} functions.
}
\keyword{internal}
