\name{ape-internal}
\alias{tree.build}
\alias{f.cherry.yule}
\alias{f.cherry.uniform}
\alias{sort.index}
\alias{nsca}
\alias{get.dna}
\alias{perm.rowscols}
\alias{mant.zstat}
\alias{lower.triang}
\alias{clado.build}
\title{Internal Ape Functions}
\description{
  Internal ape functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
