\name{treePop}
\alias{treePop}
\title{Tree Popping}
\description{
  Method for reconstructing phylogenetic trees from an object of class
  splits using tree popping.
}
\usage{
treePop(obj)
}
\arguments{
  \item{obj}{an object of class \code{"bitsplit"}.}
}
\value{
  an object of class "phylo" which displays all the splits
  in the input object.
}
\author{Andrei Popescu}
\keyword{models}
